% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotAssist.R
\name{ggplotAssist}
\alias{ggplotAssist}
\title{A shiny app for learn dplyr}
\usage{
ggplotAssist(df = NULL, viewer = "browser")
}
\arguments{
\item{df}{A tibble or a tbl_df or a data.frame to manipulate}

\item{viewer}{Specify where the gadget should be displayed. Possible choices are c("dialog","browser","pane")}
}
\value{
A manipulated tibble or NULL
}
\description{
A shiny app for learn dplyr
}
\examples{
library(tidyverse)
library(rstudioapi)
library(miniUI)
library(moonBook)
library(shinyAce)
library(ggthemes)
library(shiny)
library(stringr)
library(editData)
library(shinyWidgets)
library(gcookbook)

\dontrun{
ggplotAssist(mtcars)
result<-ggplot(iris)
cat(attr(result,"code"))
}
}
