% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotAssist.R
\name{selectizeInput3}
\alias{selectizeInput3}
\title{side-by-side selectizeInput}
\usage{
selectizeInput3(..., width = 100)
}
\arguments{
\item{...}{Further arguments to be passed to selectizeInput}

\item{width}{Input width in pixel}
}
\description{
side-by-side selectizeInput
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
  ui <- fluidPage(
         selectizeInput3("color", "color", choices=colors())
    )
    server <- function(input, output) {

    }
    shinyApp(ui, server)
}
}
