% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunction.R
\name{splitData}
\alias{splitData}
\title{Elongate data.frame with column split by comma}
\usage{
splitData(df, colname)
}
\arguments{
\item{df}{a data.frame}

\item{colname}{column name}
}
\value{
An elongated data.frame
}
\description{
Elongate data.frame with column split by comma
}
