% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunction.R
\name{textFunction}
\alias{textFunction}
\title{Server function of textFunction shiny module}
\usage{
textFunction(input, output, session, argList = reactive(argList),
  editCode = reactive(TRUE), settingData = reactive(NULL))
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{argList}{A list containing options}

\item{editCode}{Logical. Wheter or not edit initial R code}

\item{settingData}{A data.frame contains information about functions}
}
\description{
Server function of textFunction shiny module
}
