% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunction.R
\name{textInput4}
\alias{textInput4}
\title{Create side-by side textInput with disabled spell check}
\usage{
textInput4(inputId, label, value = "", width = 100, bg = NULL, ...)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{width}{The width of the input in pixel}

\item{bg}{backgroung color}

\item{...}{arguments to be passed to textInput}
}
\description{
Create side-by side textInput with disabled spell check
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
     ui <- fluidPage(
          textInput4("id", "id", ""),
          textInput4("name","name","")
     )
     server <- function(input, output) {
          
     }
     shinyApp(ui, server)
}  
}
