% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textFunction.R
\name{uiOutput3}
\alias{uiOutput3}
\title{Create side-by side uiOutput}
\usage{
uiOutput3(...)
}
\arguments{
\item{...}{arguments to be passed to uiOutput}
}
\description{
Create side-by side uiOutput
}
\examples{
library(shiny)
# Only run examples in interactive R sessions
if (interactive()) {
     ui <- fluidPage(
          textInput4("name","name",""),
          uiOutput3("test")
     )
     server <- function(input, output) {
          
     }
     shinyApp(ui, server)
}  
}
